debugging/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "devpi-debugging"
version = "0.3.0"
description = "devpi-debugging: a debugging view for devpi-server"
dependencies = [
    'devpi-common',
    'devpi-server',
    'devpi-web',
    'pyramid-debugtoolbar',
]
license = {file = "LICENSE"}
requires-python = ">=3.9"


[project.entry-points.devpi_server]
devpi-debugging = "devpi_debugging.main"


[tool.ruff]
extend = "../pyproject.toml"
target-version = "py39"


[tool.ruff.format]
exclude = [
    "devpi_debugging/main.py",
    "devpi_debugging/views.py",
]


[tool.ruff.lint]
extend-ignore = [
    "B007", # maybe cleanup later
    "BLE001", # cleanup soon - blind exception
    "PERF401", # maybe cleanup later - list comprehension
    "SIM102", # maybe cleanup later
]


[tool.ruff.lint.extend-per-file-ignores]
"devpi_debugging/main.py" = [
    "I001", # maybe cleanup later - import block unsorted
]


[tool.setuptools]
packages = [
    'devpi_debugging',
]
